Move usage string to internal to fix `go run`
